card buttons was moved to separate element
